+++ internal/client/backup_spp_server.go
@@ -47,7 +47,7 @@ func (c *BackupSPPServerClient) Read(ctx context.Context, id string) (*BackupSPP
 		return nil, err
 	}
 	defer closeResponseBody(resp)
-	found, err := requireNotFoundOrOK(resp, 403)
+	found, err := requireNotFoundOrOK(resp, 404)
 	if err != nil || !found {
 		return nil, err
 	}